one small tweak to the last commit.
git-svn-id: http://code.icecube.wisc.edu/svn/projects/dataclasses/trunk@86641 16731396-06f5-0310-8873-f7f720988828

+++ private/dataclasses/physics/I3MCHit.cxx
@@ -37,7 +37,7 @@ void I3MCHit::load (Archive &ar, const unsigned version)
   }else{
     double weight;
     ar & make_nvp("Weight", weight);
-    npe_ = static_cast<unsigned>(weight);
+    npe_ = static_cast<uint64_t>(weight);
   }
   ar & make_nvp("ParticleID", particleID_);
   if(version>0)
